README: Use syntax highlighter that allows comments
Use the `jsonc` instead of the `json` syntax highlighter in the README
to avoid the comments being marked as error.

The GitHub syntax highlighter seems to have parsing issues with the use
of comment tokens `/**` and `*/` within strings: it thinks those are
JSON comments, even though they are normal strings. Converting the JSON
fragment into a full JSON document with curly braces at the beginning
and end brings the parser back on track and gets us the right syntax
highlighting.